+++ src/renderer/components/Welcome/Welcome.stories.tsx
@@ -10,5 +10,9 @@ const meta: Meta = {
 export default meta;
 
 export const Basic = () => {
-	return <Welcome onDismiss={() => {}} />;
+	return <Welcome appVersion="mock" onDismiss={() => {}} />;
+};
+
+export const AlphaWarning = () => {
+	return <Welcome appVersion="0.0.0" onDismiss={() => {}} />;
 };