+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+# Alpha warning
+
+Until AMMiSTer reaches 1.0.0, it is considerd "alpha". More features and improvements are coming. And there may be bugs I haven't found yet.
+
+# Version 0.0.2
+
+- [All] Welcome has a warning about being an alpha version
+
 # Version 0.0.1
 
 - [Windows] Bug Fix: prevented app from crashing at startup